+++ database/migrations/2025_02_07_094559_create_station_vehicle_table.php
@@ -6,24 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('station_vehicle', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('stations_id')->constrained('stations')->onDelete('cascade');
-            $table->foreignId('vehicles_id')->constrained('vehicles')->onDselete('cascade');
-            $table->bigInteger('total_collected');
-            $table->bigInteger('pass_count');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('vehicles')) {
+            Schema::create('station_vehicle', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('stations_id')->constrained('stations')->onDelete('cascade');
+                $table->foreignId('vehicles_id')->constrained('vehicles')->onDselete('cascade');
+                $table->bigInteger('total_collected');
+                $table->bigInteger('pass_count');
+                $table->timestamps();
+            });
+        }
     }
-
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('station_vehicle');
